Make general runner and add day 3 2024
Instead of downloading the file and putting it in the right place I now
simply add the session token from the browser in a .env file and the day's
input gets downloaded if not present in the inputs directory.

This commit also contains the solution to part 2 of day 3 for 2024. It's solved
using a regex approach which I personally think is crap but it's the quickest way
to the goal for a generally uninteresting problem.