Add day 7 of 2024
This has to be the easiest part2 ever seen. Now, this is some terrible python
code but it gets the job done. I'm just glad the second part didn't ask for
the number of correct permutations for all numbers cause that would take some
more brain power than this approach.